+++ evtloop.c
@@ -305,8 +305,8 @@ auef_readable(UNUSED int fd, UNUSED void *udata) {
 			break;
 		}
 		assert(ev.subject_present);
-		assert(ev.process_present);
-		hackmon_ptrace(&ev.tv, &ev.subject, &ev.process);
+		assert(ev.args[1].present); /* pid, can be -1 */
+		hackmon_ptrace(&ev.tv, &ev.subject, (pid_t)ev.args[1].value);
 		break;
 
 	/*
@@ -550,9 +550,11 @@ siginfo_arrived(UNUSED int sig, UNUSED void *udata) {
 	fprintf(stderr, "hackmon "
 	                "recvd:%"PRIu64" "
 	                "procd:%"PRIu64" "
+	                "pidmiss:%"PRIu64" "
 	                "oom:%"PRIu64"\n",
 	                st.hm.receiveds,
 	                st.hm.processeds,
+	                st.hm.pidmiss,
 	                st.hm.ooms);
 
 	fprintf(stderr, "filemon "