+++ hackmon.h
@@ -23,6 +23,7 @@
 typedef struct {
 	uint64_t receiveds;
 	uint64_t processeds;
+	uint64_t pidmiss;
 	uint64_t ooms;
 } hackmon_stat_t;
 
@@ -37,8 +38,8 @@ typedef struct {
 
 void hackmon_taskforpid(struct timespec *tv, audit_proc_t *, audit_proc_t *)
      NONNULL(1,2,3);
-void hackmon_ptrace(struct timespec *tv, audit_proc_t *, audit_proc_t *)
-     NONNULL(1,2,3);
+void hackmon_ptrace(struct timespec *tv, audit_proc_t *, pid_t)
+     NONNULL(1,2);
 
 void hackmon_init(config_t *) NONNULL(1);
 void hackmon_fini(void);